Move timeline controller base name and package
